moved implementation to hpp file
